Move go_annotations to data and update tests
The QuickGO tests are now updated to reflect the new, more correct
location for go_annotations data structures.
