+++ rnacentral_pipeline/databases/data/go_annotations.py
